+++ Source/Classes/View/CardView/Front/SmallFrontView.swift
@@ -38,6 +38,8 @@ public class SmallFrontView: CardView {
     
     @IBOutlet weak var paymentMethodImageWidthConstraint: NSLayoutConstraint!
     
+    @IBOutlet weak var cardBalanceContainer: CardBalance!
+
     override func setupUI(_ cardUI: CardUI) {
         super.setupUI(cardUI)
         layer.cornerRadius = CardCornerRadiusManager.getCornerRadius(from: .large)
@@ -205,6 +207,13 @@ public class SmallFrontView: CardView {
             self.customView = nil
         }
         
+    }
+    
+    override func addCardBalance(_ model: CardBalanceModel, _ showBalance: Bool, _ delegate: CardBalanceDelegate) {
+        cardBalanceContainer.model = model
+        cardBalanceContainer.showBalance = showBalance
+        cardBalanceContainer.delegate = delegate
+        cardBalanceContainer.render()
     }
 }