+++ Source/Classes/View/CardView/GenericView.swift
@@ -8,7 +8,7 @@
 import UIKit
 
 public class GenericView: UIView, BasicCard  {
-    
+
     @IBOutlet weak var imageContainerView: UIView!
     @IBOutlet weak var imageView: UIImageView!
     @IBOutlet weak var titleLabel: UILabel!
@@ -193,4 +193,6 @@ public class GenericView: UIView, BasicCard  {
     func setupAnimated(_ cardUI: CardUI) {
         
     }
+    
+    func addCardBalance(_ model: CardBalanceModel, _ showBalance: Bool, _ delegate: CardBalanceDelegate) {}
 }