+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [Unreleased]
+
+## Added
+- Card Balance Component
+- TagBottom component
+
 ## [Release (1.10.0)](https://github.com/mercadolibre/meli-card-drawer-ios/releases/tag/1.10.0)
 ### 🚀 Feature 🚀
 - Added CustomStringConvertible conformance to MLCardDrawerTypeV3